Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
kubernetes>=31.0.0
pydantic>=2.11
pydantic-settings>=2.9
python-dotenv>=1.0.1
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
asyncpg>=0.30.0
httpx>=0.27.2
pyyaml>=6.0.2
ruff>=0.7.0
black>=24.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
itsdangerous>=2.1.2
